Add logger.py
